Correcciones

+++ appDisney/src/main/java/com/appDisney/repositorio/RepositorioPersonaje.java
@@ -2,6 +2,8 @@ package com.appDisney.repositorio;
 
 
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -14,7 +16,8 @@ public interface RepositorioPersonaje extends JpaRepository<Personaje, Long>{
 
 	ArrayList<Personaje> findByNombre(String nombre);
 
-	ArrayList<Personaje> findByEdad(long edad);
+	ArrayList<Personaje> findByEdad(int edad);
+	
 
 	
 }
